Data.H3.Svg

Description

Render Shapes to SVG using the svg-builder package.

Documentation

renderShape :: Shape Text (Pixel Double, Pixel Double) -> Element

Render a Shape as an SVG Element

renderSvg :: ViewBoxMode -> ((Extent Double, Extent Double) -> Shape Text (Pixel Double, Pixel Double)) -> (Extent Double, Extent Double) -> Element

Render a shape to an SVG document using the provided dimensions, potentially adding a viewbox.

renderSvg' :: ((Extent Double, Extent Double) -> Shape Text (Pixel Double, Pixel Double)) -> (Extent Double, Extent Double) -> Element

Render a shape to an SVG document using the provided dimensions

data ViewBoxMode

Whether to include a viewbox attribute in the generated SVG

Constructors

AddViewBox

Fields

vbScaleFactor :: Double

NoViewBox